Second iteration

+++ src/main/java/unq/tip/travelingtips/model/Marker.java
@@ -11,7 +11,7 @@ public class Marker implements Serializable {
 
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String latitude;
     private String longitude;
@@ -34,11 +34,11 @@ public class Marker implements Serializable {
         return serialVersionUID;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -81,4 +81,5 @@ public class Marker implements Serializable {
     public void setRating(Integer rating) {
         this.rating = rating;
     }
+
 }
